[WIP] pausing settings to work on reorganizing stages
